+++ packages/fastboard-ui/src/components/PageControl/PageControl.svelte
@@ -28,7 +28,8 @@
   const name = "fastboard-page-control";
 
   $: writable = app?.writable;
-  $: disabled = !$writable;
+  $: phase = app?.phase;
+  $: disabled = !($writable && $phase === "connected");
   $: t = i18n[language];
 
   let type: IconType;