+++ packages/fastboard-ui/src/components/ZoomControl/ZoomControl.svelte
@@ -49,7 +49,8 @@
   const name = "fastboard-zoom-control";
 
   $: writable = app?.writable;
-  $: disabled = !$writable;
+  $: phase = app?.phase;
+  $: disabled = !($writable && $phase === "connected");
   $: t = i18n[language];
 
   let type: IconType;